beets-465 Moved functionality to Weather Widget
