+++ client/src/components/NavBar/NavBar.js
@@ -14,7 +14,7 @@ import checkIn from "../../assets/checkIn.png";
 import operators from "../../assets/operators.png";
 import logOut from "../../assets/logOut.png";
 import signIn from "../../assets/signIn.png";
-import Card from "../WeatherWidget/Card";
+import WeatherWidget from "../WeatherWidget/WeatherWidget";
 import "./NavBar.css";
 
 function NavBar(props) {
@@ -37,9 +37,6 @@ function NavBar(props) {
         width="200"
       ></img>
       <ul>
-        <li style={{ color: "whitesmoke" }} >
-          <Card crossorigin apiKey={"b3cd383e1a41099de4513c032475c2ea"}/>
-        </li>
         <li>
           <DarkModeSwitch onChange={handleClick} />
         </li>
@@ -132,6 +129,13 @@ function NavBar(props) {
             </AccessControl>
           </Modal.Body>
         </Modal>
+
+        <li className="weather">
+          <WeatherWidget
+            crossorigin
+            apiKey={"b3cd383e1a41099de4513c032475c2ea"}
+          />
+        </li>
       </ul>
     </nav>
   );
