+++ client/src/components/WeatherWidget/WeatherWidget.css
@@ -1,0 +1,4 @@
+.WeatherWidget {
+    color: whitesmoke;
+    font-size: .9em;
+}
